% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_circos.R
\name{calc_circos}
\alias{calc_circos}
\alias{circos}
\title{Calculate linkage disequilibrium circos}
\usage{
calc_circos(geno012, snpinfo, method = "RMI", cores = 1, threshold = 0.2)

circos(m, snpinfo, threshold = 0.2)
}
\arguments{
\item{geno012}{a genetype matrix coded by 0,1,2, NA(NA for missing values)}

\item{snpinfo}{a matrix or data.frame consists of three columns(chr,pos,id)}

\item{method}{linkage disequilibrium measures}

\item{cores}{The number of threads}

\item{threshold}{the minimum threshold of preserve pairwise calculations}

\item{m}{a upper.tri matrix calculated by linkage disequilibrium measures}
}
\description{
Calculate linkage disequilibrium circos
}
