% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_decay.R
\name{calc_decay}
\alias{calc_decay}
\alias{decay}
\title{Calculate linkage disequilibrium decay}
\usage{
calc_decay(geno012, snpinfo, method = "RMI", cores = 1, maxdist = 300)

decay(m, snpinfo, maxdist = 300)
}
\arguments{
\item{geno012}{a genetype matrix coded by 0,1,2, NA(NA for missing values)}

\item{snpinfo}{a matrix or data.frame consists of three columns(chr,pos,id)}

\item{method}{linkage disequilibrium measures}

\item{cores}{The number of threads}

\item{maxdist}{The maximum distance calculated by pairwise SNP}

\item{m}{a upper.tri matrix calculated by linkage disequilibrium measures}
}
\description{
Calculate linkage disequilibrium decay
}
