% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos.legendlink}
\alias{circos.legendlink}
\title{circos legend}
\usage{
circos.legendlink(
  x = x,
  y = y,
  legend.limits = c(min, max),
  legend.n.breaks = legend.n.breaks,
  legend.breaks = legend.breaks,
  legend.labels = legend.labels,
  legend.col = legend.col,
  legend.fontsize = legend.fontsize,
  legend.family = legend.family,
  ...
)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{legend.limits}{A numeric vector of length two providing limits of the scale}

\item{legend.n.breaks}{An integer guiding the number of major breaks.}

\item{legend.breaks}{A numeric vector of positions}

\item{legend.labels}{A vector giving labels (must be same length as legend.breaks)}

\item{legend.col}{A specification for the default plotting color}

\item{legend.fontsize}{numeric; the font size of legend}

\item{legend.family}{The font family of legend}

\item{...}{other graphical parameters}
}
\description{
circos legend
}
