% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos.linksnp}
\alias{circos.linksnp}
\title{link snp}
\usage{
circos.linksnp(
  data,
  ylim = c(0, 1),
  link.col = NULL,
  track.height = 0.05,
  legend = TRUE,
  legend.pos = c(0.9, 0.6),
  legend.border = NA,
  legend.font = 1,
  legend.labels.cex = 0.75,
  bg.border = "black",
  bg.col = NA,
  family = NULL,
  lwd = par("lwd"),
  lty = par("lty"),
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{ylim}{Range of data on y-axis}

\item{link.col}{link lines' color}

\item{track.height}{Height of the track}

\item{legend}{logical whether to show the legend}

\item{legend.pos}{legend position}

\item{legend.border}{Color for the border of the legend}

\item{legend.font}{font for the legend labels}

\item{legend.labels.cex}{Color for the border of the plotting regions}

\item{bg.border}{Color for the border of the plotting regions}

\item{bg.col}{Background color for the plotting regions}

\item{family}{font of legend label}

\item{lwd}{The line width}

\item{lty}{The line type}

\item{...}{other graphical parameters}
}
\description{
link snp
}
\examples{
NULL
}
