% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegeno.R
\name{codegeno}
\alias{codegeno}
\alias{codegeno,data.frame-method}
\alias{codegeno,matrix-method}
\title{Convert character genotype matrix to integer genotype matrix}
\usage{
codegeno(data, sep, ...)

\S4method{codegeno}{data.frame}(data, sep, ...)

\S4method{codegeno}{matrix}(data, sep, ...)
}
\arguments{
\item{data}{character numeric matrix of genotype}

\item{sep}{the separator for character genotypes matrix(i.e., "" ,"/", "|")}

\item{...}{optional arguments (ignored)}
}
\description{
convert genotypes matrix separated by ("", "/", "|", ...) to integer genotypes matrix(0, 1, 2, NA, NA for missing value)
}
\examples{
g0 <- c('T/A', NA, 'T|T', NA, 'T|A', NA, 'T|T', 'T/A','T/T', 'T/T', 
        'T/A', 'A|A', 'T/T', 'T|A', 'T/A', 'T|T', NA, 'T|A', 'T/A', NA)

g1 <- c("0/0", "1/1", "0/1", "0|0", "1|1", "1/1", "1/1", "0/1", "0/0", "0/0", 
        "1|1", "0|1", "0|0", "0|0", "1|1", "0/0", "0|0", "0|0", "1|1", "0/0") 

gt_mat <- as.matrix(data.frame(g0=g0, g1=g1))

geno012 <- codegeno(gt_mat, "/|")
}
