% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWLD.R
\name{GWLD}
\alias{GWLD}
\alias{data.frame,ANY}
\alias{matrix,ANY}
\alias{numeric,numeric}
\alias{GWLD,data.frame,ANY-method}
\alias{GWLD,matrix,ANY-method}
\alias{GWLD,numeric,numeric-method}
\title{GWLD: A package for Genome wide linkage disequilibrium calculate}
\usage{
GWLD(g1, g2, method = "RMI", cores = 1, ...)

\S4method{GWLD}{data.frame,ANY}(g1, g2, method = "RMI", cores = 1, ...)

\S4method{GWLD}{matrix,ANY}(g1, g2, method = "RMI", cores = 1, ...)

\S4method{GWLD}{numeric,numeric}(g1, g2, method = "RMI", cores = 1, ...)
}
\arguments{
\item{g1}{vector,dataframe or matrix containing genotype coded with interger(0, 1, 2, NA)}

\item{g2}{vector(ignored if g1 is a dataframe or matrix)}

\item{method}{methods}

\item{cores}{thread number}

\item{...}{optional arguments (ignored)}
}
\description{
GWLD: A package for Genome wide linkage disequilibrium calculate
}
