% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{HeatMap}
\alias{HeatMap}
\title{linkage disequilibrium heatmap}
\usage{
HeatMap(
  data,
  method = "r^2",
  SnpPosition = NULL,
  SnpName = NULL,
  cores = 1,
  color = NULL,
  showLDvalues = FALSE,
  title = "Pairwise LD",
  label.size = 10,
  family = NULL,
  ...
)
}
\arguments{
\item{data}{SNP genotype data.frame or matrix}

\item{method}{measure linkage disequilibrium method}

\item{SnpPosition}{vector}

\item{SnpName}{vector}

\item{cores}{thread number}

\item{color}{heatmap corlor}

\item{showLDvalues}{logical}

\item{title}{heatmap's title}

\item{label.size}{label's font size of heatmap}

\item{family}{Font for heatmap}

\item{...}{NULL}
}
\value{
HeatMap
}
\description{
linkage disequilibrium heatmap
}
\examples{
NULL
}
