% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD}
\alias{LD}
\alias{LD,data.frame,ANY-method}
\alias{LD,matrix,ANY-method}
\alias{LD,numeric,numeric-method}
\title{GWLD: a package for genome-wide linkage disequilibrium analysis}
\usage{
LD(g1, g2, method = "r^2", cores = 1, ...)

\S4method{LD}{data.frame,ANY}(g1, method = "r^2", cores = 1)

\S4method{LD}{matrix,ANY}(g1, method = "r^2", cores = 1)

\S4method{LD}{numeric,numeric}(g1, g2, method = "r^2", cores = 1, ...)
}
\arguments{
\item{g1}{vector,dataframe or matrix containing genotype coded with interger(0, 1, 2, NA)}

\item{g2}{vector(ignored if g1 is a dataframe or matrix)}

\item{method}{methods}

\item{cores}{thread number}

\item{...}{optional arguments (ignored)}
}
\value{
a matrix, row names are the same as column names.
}
\description{
GWLD: a package for genome-wide linkage disequilibrium analysis
}
\examples{
g0 <- c('T/A', NA, 'T|T', NA, 'T|A', NA, 'T|T', 'T/A','T/T', 'T/T',
        'T/A', 'A|A', 'T/T', 'T|A', 'T/A', 'T|T', NA, 'T|A', 'T/A', NA)

g1 <- c('C/A', 'C/A', 'C/C', 'C/A', 'C/C', 'C/A', 'C/A', 'C/A',
       'C/A', 'C/C', 'C/A', 'A/A', 'C/A', 'A/A', 'C/A', 'C/C',
       'C/A', 'C/A', 'C/A', 'A/A')

gt <- data.frame(g0, g1)
geno012 <- codegeno(gt, "/|")
LD(geno012, method="r^2", cores=1)

g2 <- recode(g0,"/|")
g3 <- recode(g1,"/")

LD(g2,g3, method="r^2")

g4 <- data.frame(g2, g3)

LD(g4, method="r^2", cores=2)

}
