% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.R
\name{MI}
\alias{MI}
\alias{MI,data.frame,ANY-method}
\alias{MI,matrix,ANY-method}
\alias{MI,numeric,numeric-method}
\title{Compute pairwise linkage disequilibrium with mutual information}
\usage{
MI(g1, g2, cores = 1, ...)

\S4method{MI}{data.frame,ANY}(g1, cores = 1)

\S4method{MI}{matrix,ANY}(g1, cores = 1)

\S4method{MI}{numeric,numeric}(g1, g2)
}
\arguments{
\item{g1}{vector,dataframe or matrix containing genotype coded with interger(0, 1, 2, NA)}

\item{g2}{vector(ignored if g1 is a dataframe or matrix)}

\item{cores}{thread number}

\item{...}{optional arguments (ignored)}
}
\value{
a matrix, row names are the same as column names.
}
\description{
Compute pairwise linkage disequilibrium with mutual information
}
\examples{
g0 <- c('T/A', NA, 'T|T', NA, 'T|A', NA, 'T|T', 'T/A','T/T', 'T/T',
        'T/A', 'A|A', 'T/T', 'T|A', 'T/A', 'T|T', NA, 'T|A', 'T/A', NA)

g1 <- c('C/A', 'C/A', 'C/C', 'C/A', 'C/C', 'C/A', 'C/A', 'C/A',
       'C/A', 'C/C', 'C/A', 'A/A', 'C/A', 'A/A', 'C/A', 'C/C',
       'C/A', 'C/A', 'C/A', 'A/A')

gt <- data.frame(g0, g1)
geno012 <- codegeno(gt, "/|")
MI(geno012, cores=2)

g2 <- recode(g0,"/|")
g3 <- recode(g1,"/")

MI(g2,g3)

g4 <- data.frame(g2, g3)

MI(g4, cores=2)
}
