% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedeal.R
\name{ped2matrix}
\alias{ped2matrix}
\title{plink format(.ped, .map) to  matrix}
\usage{
ped2matrix(FilePrefix, ...)
}
\arguments{
\item{FilePrefix}{plink format(.ped, .map) file prefix}

\item{...}{NULL}
}
\description{
plink format(.ped, .map) to  matrix
}
