% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedeal.R
\name{read.plink}
\alias{read.plink}
\title{read plink format(.bed, .bim, .fam)}
\usage{
read.plink(FilePrefix, ...)
}
\arguments{
\item{FilePrefix}{plink format(.bed, .bim, .fam) file's prefix}

\item{...}{NULL}
}
\value{
list
}
\description{
read plink format(.bed, .bim, .fam)
}
