% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedeal.R
\name{read.vcf}
\alias{read.vcf}
\title{read vcf format file}
\usage{
read.vcf(filename, genotype = "None", ...)
}
\arguments{
\item{filename}{vcf format file's name}

\item{genotype}{genotype display type, one of "None, "char", "int" and "allele"}

\item{...}{NULL}
}
\value{
matrix
}
\description{
read vcf format file
}
