% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos.ideogram}
\alias{circos.ideogram}
\title{circos snp}
\usage{
circos.ideogram(
  data,
  ylim = c(0, 1),
  cell.padding = c(0.02, 0, 0.02, 0),
  track.height = 0.1,
  cex.lab = 0.6,
  bg.border = NA,
  family = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{ylim}{Range of data on y-axis}

\item{cell.padding}{only affect current track}

\item{track.height}{Height of the track. It is the percentage to the radius of the unit circles}

\item{cex.lab}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the defaul}

\item{bg.border}{Color for the border of the plotting regions}

\item{family}{the name of a font family for drawing text}

\item{...}{other graphical parameters}
}
\description{
circos snp
}
\examples{
NULL
}
