% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stats_freq}
\alias{stats_freq}
\title{stats allele frequency
 
stats allele frequency of character genotypes vector}
\arguments{
\item{vector}{a character vector of genotypes}

\item{sep}{the separator for character genotypes(i.e., "" ,"/", "|")}
}
\value{
list
}
\description{
stats allele frequency
 
stats allele frequency of character genotypes vector
}
\examples{

g0 <- c('T/A', NA, 'T|T', NA, 'T|A', NA, 'T|T', 'T/A','T/T', 'T/T', 
        'T/A', 'A|A', 'T/T', 'T|A', 'T/A', 'T|T', NA, 'T|A', 'T/A', NA)

stats_freq(g0, "/|")
}
