% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos.linksnp}
\alias{circos.linksnp}
\title{link snp}
\usage{
circos.linksnp(
  data,
  ylim = c(0, 1),
  color.theme = "viridis",
  track.height = 0.05,
  legend.show = TRUE,
  legend.position = c(0.88, 0.75),
  legend.fontsize = 12,
  legend.limits = NULL,
  legend.breaks = NULL,
  legend.labels = NULL,
  legend.n.breaks = NULL,
  family = NULL,
  bg.border = "black",
  bg.col = NULL,
  lwd = par("lwd"),
  lty = par("lty"),
  ...
)
}
\arguments{
\item{data}{A data frame with 7 columns;Chr_1, Pos_1, ID_1, Chr_2, Pos_2, ID_2 value}

\item{ylim}{Range of data on y-axis}

\item{color.theme}{The color theme in the legend}

\item{track.height}{Height of the track}

\item{legend.show}{logical whether to show the legend}

\item{legend.position}{legend position}

\item{legend.fontsize}{the font size of legend}

\item{legend.limits}{A numeric vector of length two providing limits of the scale}

\item{legend.breaks}{A numeric vector of positions}

\item{legend.labels}{A vector giving labels (must be same length as legend.breaks)}

\item{legend.n.breaks}{An integer guiding the number of major breaks}

\item{family}{The font family of legend}

\item{bg.border}{Color for the border of the plotting regions}

\item{bg.col}{Background color for the plotting regions}

\item{lwd}{The line width}

\item{lty}{The line type}

\item{...}{other graphical parameters}
}
\description{
link snp
}
\examples{
NULL
}
